backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

transformers==4.40.0
torch>=2.1.0

pillow==10.3.0
python-multipart==0.0.9

# Optional but recommended
accelerate==0.29.3